poll: Always let the winsys add fds even if there is a zero timeout
Even if Cogl decides to set a zero timeout because there are events
queued, it still makes sense to give the winsys a chance to add file
descriptors to the list. The winsys might be relying on the list of
CoglPollFDs passed to poll_dispatch to decide whether to read from a
file descriptor and that should happen even if Cogl also woke up the
main loop because the event queue isn't empty.

Reviewed-by: Robert Bragg <robert@linux.intel.com>

(cherry picked from commit 6d2f3bc4913d0f1570c09e3714ac8fe2dbfc7a03)